add usart alternate port support
Add ~1.6K if selected to provide STM bootblock serial protocol flash
compatible support as defined in AN3155, which interfaces to a
stm32flash tool.  Default is 921600 Baud, 8 bits data, even parity,
2 bits stop.

Config needs to set DFU_USART or DFU_UART to the configured UART
number.  Further refinements to the exact ports being used, should
the defaults in usart.c not be sufficient, can be facilitated by
defining the following to GPIOx_BASE and GPIO_IN_x respectively:

    USARTx_TX_GPIO_PORT  USARTx_TX_GPIO_PIN
    USARTx_RX_GPIO_PORT  USARTx_RX_GPIO_PIN
    USARTx_RTS_GPIO_PORT USARTx_RTS_GPIO_PIN
    USARTx_CTS_GPIO_PORT USARTx_CTS_GPIO_PIN
    USARTx_DCD_GPIO_PORT USARTx_DCD_GPIO_PIN

The DCD signal is not hardware supported, per se, but can be used to
select a pin that gets pulled up or down when serial port is active,
and pulled down or up when serial port is inactive, to complete the
serial port signature.

Signed-off-by: Mark Salyzyn <mark.salyzyn@gmail.com>
Test: diff -ub src/usart.c ../../libdmt/src/dfu.cpp